+++ src/exception/ConditionViolationException.js
@@ -48,7 +48,8 @@ ConditionViolationException.prototype._getTemplateVariables = function () {
  * @inheritdoc
  */
 ConditionViolationException.prototype.toString = function () {
-    return util.format('Condition violation at "%s": %s', this.pathStr, this.message);
+    var location = this.path.length === 0 ? 'input value' : '"' + this.pathStr + '"';
+    return util.format('Condition violation at %s: %s', location, this.message);
 };
 
 module.exports = ConditionViolationException;